Move to production src folder

rh-pre-commit.version: 2.0.1
rh-pre-commit.check-secrets: ENABLED
