+++ kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/api/common/status/KafkaServiceStatus.java
@@ -7,7 +7,7 @@
 package io.kroxylicious.kubernetes.api.common.status;
 
 @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)
-@com.fasterxml.jackson.annotation.JsonPropertyOrder({"conditions","observedGeneration"})
+@com.fasterxml.jackson.annotation.JsonPropertyOrder({ "conditions", "observedGeneration" })
 @com.fasterxml.jackson.databind.annotation.JsonDeserialize(using = com.fasterxml.jackson.databind.JsonDeserializer.None.class)
 @javax.annotation.processing.Generated("io.fabric8.java.generator.CRGeneratorRunner")
 @lombok.ToString()
@@ -25,4 +25,3 @@ package io.kroxylicious.kubernetes.api.common.status;
 public class KafkaServiceStatus extends CommonStatus {
 
 }
-
