+++ kroxylicious-operator/src/test/java/io/kroxylicious/kubernetes/operator/assertj/OperatorAssertions.java
@@ -7,9 +7,11 @@
 package io.kroxylicious.kubernetes.operator.assertj;
 
 import io.kroxylicious.kubernetes.api.common.Condition;
+import io.kroxylicious.kubernetes.api.common.status.KafkaServiceStatus;
 import io.kroxylicious.kubernetes.api.v1alpha1.KafkaProxyStatus;
-import io.kroxylicious.kubernetes.api.v1alpha1.KafkaServiceStatus;
 import io.kroxylicious.kubernetes.api.v1alpha1.kafkaproxystatus.Clusters;
+
+;
 
 public class OperatorAssertions {
     public static KafkaProxyStatusAssert assertThat(KafkaProxyStatus actual) {
